Move evaluator-related things to their respective files
