+++ spsp/__init__.py
@@ -1,18 +1,5 @@
-from .tokenizer import Tokenizer
-from . import errors
-from . import Token
-from .parser import parse
-from .scope import Scope
-from .evaluation import evaluate
-
-__all__ = [
-    'Tokenizer',
-    'Token',
-
-    'parse',
-
-    'errors',
-
-    'Scope',
-    'evaluate'
-]
+# Import for side effects
+# Importing these scripts will fun the decorators which
+# populate evaluator mappings
+from .evaluation_rule_definitions import *
+from .special_form_definitions import *
